% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allFunctions.R
\name{allFunctions}
\alias{allFunctions}
\title{allFunctions}
\usage{
allFunctions()
}
\value{
character vector of functions names
}
\description{
Returns the names of all functions currently available.
}
\examples{
allFunctions()
}
\references{
https://stackoverflow.com/questions/4267744/is-there-a-way-to-get-a-vector-with-the-name-of-all-functions-that-one-could-use
}
