% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMsg.R
\name{getMsg}
\alias{getMsg}
\title{getMsg}
\usage{
getMsg(
  ...,
  .domain = getOption("stranslate.domain"),
  .lang = getOption("stranslate.lang")
)
}
\arguments{
\item{...}{parameter(s) given to the function}

\item{.domain}{character: domain namesd (default: \code{getOption("stranslate.domain")})}

\item{.lang}{character: language to use (default: \code{getOption("stranslate.lang")})}
}
\value{
the (translated) message
}
\description{
Returns a message. The first parameter must be the key to the message. For details read the vignette \code{vignette("stranslate")}.
}
\examples{
1+1

}
