% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listMsg.R
\name{listMsg}
\alias{listMsg}
\title{listMsg}
\usage{
listMsg(.domain = NULL)
}
\arguments{
\item{.domain}{character: domains(s) to list (default: \code{NULL})}
}
\value{
nothing
}
\description{
Lists all keys in all languages in a specific domain or all domains (default).
}
\examples{
listMsg("stranslate")
listMsg()
}
