% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMsg.R
\name{loadMsg}
\alias{loadMsg}
\title{loadMsg}
\usage{
loadMsg(
  files,
  .domain = getOption("stranslate.domain"),
  .lang = "en",
  .silent = TRUE
)
}
\arguments{
\item{files}{character: names of text file(s) with translated messages}

\item{.domain}{character: domain namesd (default: \code{getOption("stranslate.domain")})}

\item{.lang}{character: default language (default: 'en')}

\item{.silent}{logical: should during the loading process some output shown and tests after loading run}
}
\value{
invisibly the current message environment
}
\description{
Loads translated messages from one (or more) file into package environment. For details read the vignette \code{vignette("stranslate")}.
}
\examples{
1+1
}
