% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setLang.R
\name{setLang}
\alias{setLang}
\title{setLang}
\usage{
setLang(.lang = "en", .domain = getOption("stranslate.domain"))
}
\arguments{
\item{.lang}{character: language (default: \code{"en"})}

\item{.domain}{character: domain (default: \code{getOption("stranslate.domain")})}
}
\value{
invisibly the language set
}
\description{
Sets the language and the domain.
}
\examples{
setLang('tlh')  # use klingon now ;)
}
