% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMsg.R
\name{setMsg}
\alias{setMsg}
\title{setMsg}
\usage{
setMsg(..., .silent = TRUE)
}
\arguments{
\item{...}{aaa}

\item{.silent}{logical: should during the loading process some output shown and tests after loading run}
}
\value{
aaa
}
\description{
setMsg
}
\examples{
1+1
}
