% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findText.R
\name{findText}
\alias{findText}
\title{findText}
\usage{
findText(files, pattern = "\\\\w+{2,}")
}
\arguments{
\item{files}{character: names of R file(s)}

\item{pattern}{character: pattern to match a string (default: \code{"\\\\w+{2,}"})}
}
\value{
a text matrix with strings and file with line numbers
}
\description{
Finds strings matching \code{pattern} in \code{files}; for details for matching see \code{\link[base:grep]{base::grepl()}}.
}
\examples{
1+1
}
